Show when there is a filter for the attribute and value or not
